% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_test_results.R
\name{add_test_results}
\alias{add_test_results}
\title{Calculate p-values (both per_variable and pooled) and identify important
variables when using FWER, FDR or no p-value adjustment}
\usage{
add_test_results(
  vimpermsim,
  alpha = 0.05,
  init_num_vars,
  to_show = c("FWER", "FDR", "unadjusted")
)
}
\arguments{
\item{vimpermsim}{list, an output from \code{vim_perm_sim()} function}

\item{alpha}{numeric, the significance level, must be between 0 and 1}

\item{init_num_vars}{numeric, the number of covariates originally included in
the data, required to correctly apply the Benjamini-Hochberg and Holm
p-value correction}

\item{to_show}{character, one of "FWER", "FDR" or "unadjusted". If
"unadjusted" is selected, then the final output will contain only raw,
unadjusted p-values together with the decision. If "FDR" is selected, the
final output will include both unadjusted and Benjamini-Hochberg adjusted
p-values along with the decision. The last option "FWER" includes
unadjusted, Benjamini-Hochberg and Holm adjusted p-values together with the
decision. Default is "FWER".}
}
\value{
list, contains the results stored in the output of the
\code{vim_perm_sim()} function Additionally, \verb{$test_results$per_variable} and
\code{out_add$test_results$pooled} show which variables were found to be
important when applying FWER, FDR and no p-value adjustments and when using
"per_variable" and "pooled" p-values respectively.
}
\description{
Calculate p-values (both per_variable and pooled) and identify important
variables when using FWER, FDR or no p-value adjustment
}
