% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vimps.R
\name{plot_vimps}
\alias{plot_vimps}
\title{Box plots of VIMPs obtained during the simulation along with unadjusted, FDR
and FWER adjusted p-values.}
\usage{
plot_vimps(
  wrapper_object,
  pooled = TRUE,
  filter_vars = NULL,
  text_size = 3,
  ...
)
}
\arguments{
\item{wrapper_object}{the output of the function \code{vim_perm_sim_wrapper()}}

\item{pooled}{logical, argument indicating whether to display results
according to pooled or per_variable p-values}

\item{filter_vars}{parameter indicating the number of variables to plot, the
default is \code{NULL}, which means that all variables considered in the last
step of the procedure are plotted}

\item{text_size}{numeric, parameter controlling the size of the printed
p-values on the plot, default 3}

\item{...}{other options used to control the appearance of the output plot}
}
\value{
ggplot object
}
\description{
Box plots of VIMPs obtained during the simulation along with unadjusted, FDR
and FWER adjusted p-values.
}
\examples{
NULL
}
