% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.shadow_vimp.R
\name{print.shadow_vimp}
\alias{print.shadow_vimp}
\title{Print shadow_vimp results}
\usage{
\method{print}{shadow_vimp}(x, ...)
}
\arguments{
\item{x}{Object of class 'shadow_vimp'}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Custom print function to display the key elements of the \code{shadow_vimp()}
results.
}
\seealso{
\code{\link{shadow_vimp}}
}
