% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowVIMP-package.R
\docType{package}
\name{shadowVIMP-package}
\alias{shadowVIMP}
\alias{shadowVIMP-package}
\title{shadowVIMP: Covariate Selection Based on VIMP Permutation-Like Testing}
\description{
A statistical method for reducing the number of covariates in an analysis by evaluating Variable Importance Measures (VIMPs) derived from the Random Forest algorithm. It performs statistical tests on the VIMPs and outputs whether the covariate is significant along with the p-values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Staburo/shadowVIMP}
  \item Report bugs at \url{https://github.com/Staburo/shadowVIMP/issues}
}

}
\author{
\strong{Maintainer}: Tim Mueller \email{mueller@staburo.de}

Authors:
\itemize{
  \item Oktawia Miluch \email{oktawia.miluch@staburo.de}
}

Other contributors:
\itemize{
  \item Staburo GmbH [copyright holder, funder]
}

}
\keyword{internal}
