% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vimps.R
\name{plot_vimps}
\alias{plot_vimps}
\title{Box plot of VIMPs and corresponding p-values}
\usage{
plot_vimps(
  shadow_vimp_out,
  pooled = TRUE,
  filter_vars = NULL,
  p_val_labels = TRUE,
  text_size = 4,
  legend.position = c("right", "left", "top", "bottom", "none"),
  category_colors = c(`FWER conf.` = "#DD5129FF", `FDR conf.` = "#0F7BA2FF",
    `Unadjusted conf.` = "#43B284FF", `Not significant` = "#898E9FFF"),
  ...
)
}
\arguments{
\item{shadow_vimp_out}{Object of the class "shadow_vimp", the output of the
function \code{shadow_vimp()}.}

\item{pooled}{Boolean
\itemize{
\item \code{TRUE} - passed \code{shadow_vimp_out} contains pooled p-values.
\item \code{FALSE} - passed \code{shadow_vimp_out} contains per variable p-values.
}}

\item{filter_vars}{Numeric, the number of variables to plot. The default is
\code{NULL}, which means that all variables considered in the last step of the
procedure (and included in the \code{ shadow_vimp_out}) will be plotted.}

\item{p_val_labels}{Boolean, controls whether the p-value labels should be
printed on the plot, default \code{TRUE}.}

\item{text_size}{Numeric, parameter that controls the size of the printed
p-values on the plot, default is 4.}

\item{legend.position}{Character, one of "right", "left", "top", "bottom" or
"none". Argument specifying the position of the legend.}

\item{category_colors}{Character of length 4, contains color assignment for
each of four possible outcomes: variable not significant, confirmed by
unadjusted, FDR and FWER adjusted p-value. The default colors are color
blind friendly.}

\item{...}{Other options used to control the appearance of the output plot.}
}
\value{
ggplot object
}
\description{
Box plot of variable importance measures obtained from the simulation, along
with unadjusted, FDR, and FWER adjusted p-values indicating whether a given
variable is informative.
}
\examples{
data(mtcars)

# When working with real data, increase the value of the `niters` and `num.trees`
# parameters to obtain trustworthy results.
\donttest{
# Pooled p-values
out_pooled <- shadow_vimp(
  data = mtcars, outcome_var = "vs",
  niters = c(10, 20, 30), num.trees = 30
)

# The following 2 lines of code produce identical plots
plot_vimps(shadow_vimp_out = out_pooled, pooled = TRUE, text_size = 4)
plot_vimps(shadow_vimp_out = out_pooled, text_size = 4)

# Plot only 3 covariates
plot_vimps(shadow_vimp_out = out_pooled, filter_vars = 3)

#' # Do not display p-values on the plot
plot_vimps(shadow_vimp_out = out_pooled, p_val_labels = FALSE)

# Change the size of displayed p-values
plot_vimps(shadow_vimp_out = out_pooled, text_size = 6)

# Per variable p-values
out_per_var <- shadow_vimp(
  data = mtcars, outcome_var = "vs",
  niters = c(10, 20, 30), num.trees = 30, method = "per_variable"
)

# Set pooled to `FALSE`, otherwise the function will throw an error.
plot_vimps(shadow_vimp_out = out_per_var, pooled = FALSE)
}
}
