% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vimps.R
\name{plot_vimps}
\alias{plot_vimps}
\title{Box plot of VIMPs and corresponding p-values}
\usage{
plot_vimps(
  wrapper_object,
  pooled = TRUE,
  filter_vars = NULL,
  text_size = 3,
  ...
)
}
\arguments{
\item{wrapper_object}{List, the output of the function
\code{vim_perm_sim_wrapper()}.}

\item{pooled}{Boolean
\itemize{
\item \code{TRUE} - passed \code{wrapper_object} contains pooled p-values.
\item \code{FALSE} - passed \code{wrapper_object} contains per variable p-values.
}}

\item{filter_vars}{Numeric, the number of variables to plot. The default is
\code{NULL}, which means that all variables considered in the last step of the
procedure (and included in the \code{ wrapper_object}) will be plotted.}

\item{text_size}{Numeric, parameter that controls the size of the printed
p-values on the plot, default is 3.}

\item{...}{Other options used to control the appearance of the output plot.}
}
\value{
ggplot object
}
\description{
Box plot of variable importance measures obtained from the simulation, along
with unadjusted, FDR, and FWER adjusted p-values indicating whether a given
variable is informative.
}
\examples{
data(mtcars)

# When working with real data, increase the value of the nsims and num.trees
# parameters to obtain trustworthy results.
# Pooled p-values
out_pooled <- vim_perm_sim_wrapper(
  entire_data = mtcars, outcome_var = "vs",
  nsims = c(10, 20, 30), num.trees = 30
)

# The following 2 lines of code produce identical plots
plot_vimps(wrapper_object = out_pooled, pooled = TRUE, text_size = 4)
plot_vimps(wrapper_object = out_pooled, text_size = 4)

# Plot only 3 covariates
plot_vimps(wrapper_object = out_pooled, filter_vars = 3)

# Per variable p-values
out_per_var <- vim_perm_sim_wrapper(
  entire_data = mtcars, outcome_var = "vs",
  nsims = c(10, 20, 30), num.trees = 30, method = "per_variable"
)

# Set pooled to `FALSE`, otherwise the function will throw an error.
plot_vimps(wrapper_object = out_per_var, pooled = FALSE, text_size = 4)
}
