% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowVIMP-package.R
\docType{package}
\name{shadowVIMP-package}
\alias{shadowVIMP}
\alias{shadowVIMP-package}
\title{shadowVIMP: Covariate Selection Based on VIMP Permutation-Like Testing}
\description{
A statistical tool for reducing the number of covariates in an analysis by evaluating Variable Importance Measures (VIMPs) derived from the Random Forest algorithm. It performs a statistical tests on the VIMPs and outputs a decision as to whether the covariate is significant along with the raw, FWER and FDR corrected p-values.
}
\author{
\strong{Maintainer}: Oktawia Miluch \email{oktawia.miluch@staburo.de}

Authors:
\itemize{
  \item Tim Mueller \email{mueller@staburo.de}
}

Other contributors:
\itemize{
  \item Staburo GmbH [copyright holder, funder]
}

}
\keyword{internal}
