% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowVIMP-package.R
\docType{package}
\name{shadowVIMP-package}
\alias{shadowVIMP}
\alias{shadowVIMP-package}
\title{shadowVIMP: Covariate Selection Based on VIMP Permutation-Like Testing}
\description{
A statistical method for reducing the number of covariates in an analysis by evaluating Variable Importance Measures (VIMPs) derived from the Random Forest algorithm. The method performs statistical tests on VIMPs using permuted shadow variables to compute p-values, while preserving the correlation structure among predictors. It supports multiple testing adjustments and outputs significance flags for each covariate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OktawiaStaburo/shadowVIMP}
  \item Report bugs at \url{https://github.com/OktawiaStaburo/shadowVIMP/issues}
}

}
\author{
\strong{Maintainer}: Oktawia Miluch \email{oktawia.miluch@staburo.de}

Authors:
\itemize{
  \item Tim Mueller \email{mueller@staburo.de}
}

Other contributors:
\itemize{
  \item Staburo GmbH [copyright holder, funder]
}

}
\keyword{internal}
