% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim_perm_sim.R
\name{vim_perm_sim}
\alias{vim_perm_sim}
\title{Calculate variable importance of original values of predictors and its
row-wise permuted shadows}
\usage{
vim_perm_sim(
  entire_data,
  outcome_var,
  nsim = 100,
  importance = "permutation",
  num.threads = NULL,
  write.forest = F,
  num.trees = max(2 * (ncol(entire_data) - 1), 10000),
  data_name = NULL,
  num_cores_parallel = NULL,
  ...
)
}
\arguments{
\item{entire_data}{data frame}

\item{outcome_var}{character, name of the column containing the outcome
variable}

\item{nsim}{numeric, number of permutations, by default 100}

\item{importance}{character, variable importance mode passed directly to
\code{ranger::ranger()}, default is "permutation"}

\item{num.threads}{numeric, number of threads, default is number of CPUs
available}

\item{write.forest}{logical, indicator whether to save \code{ranger.forest}
object, default is \code{FALSE}}

\item{num.trees}{numeric, number of trees}

\item{num_cores_parallel}{numeric greater than 0 and and less than or equal to the
number of cores available on your computer (\code{detectCores()}). This
parameter specifies the number of cores to use to create a cluster by
calling \code{parallel::makeCluster(num_cores_parallel)}. The default is \code{NULL}, which
means that sequential computation is used.}

\item{...}{additional parameters passed directly to \code{ranger::ranger()} or
\code{party::cforest()}}

\item{df_name}{character, name of the object passed as \code{entire_data}. In
\code{vim_perm_sim_wrapper()} set automatically}
}
\value{
List where:
\enumerate{
\item \code{vim_simulated} stores \code{n_sim} variable importances calculated based
on the original and row-wise permuted values of the predictors,
\item \code{controls} stores used control parameters, by default - number of
permutations \code{n_sim}.
}
}
\description{
Dependencies: dplyr, ranger, magrittr, parallel, foreach, doSNOW/doParallel, doRNG
}
